MNumber -> CNumber
