+++ main/src/types/number/index.ts
@@ -1,3 +1,2 @@
 export * from "./counter";
 export { CNumberEvent, CNumberEventsRecord, CNumber } from "./number";
-export { MNumber } from "./multi_number";
