--- main/src/types/number/multi_number.proto
@@ -1,3 +1,0 @@
-message MNumberComponentMessage {
-    required double arg = 1;
-}
